fix(smb/replay): gate LOCK replay cache on durable/multichannel
LockSequence checks now mirror MS-SMB2 §3.3.5.14 / Samba: cache is
consulted only when Open.IsDurable is true or the connection
negotiated SMB2_GLOBAL_CAP_MULTI_CHANNEL. Without the gate, a basic
LOCK-stacking client (smbtorture lock.valid-request) sees its
second same-sequence LOCK short-circuit out of stacking and a
later UNLOCK fails RANGE_NOT_LOCKED.

Also fixes the LockSequence bit layout to match the spec: low 4
bits = LockSequenceNumber, upper 28 bits = LockSequenceIndex
(bucket; 1-based, capped at 64 per Samba SMB2_LOCK_SEQUENCE_ARRAY_SIZE).

+++ internal/adapter/smb/v2/handlers/lock_async.go
@@ -41,8 +41,8 @@ func (h *Handler) parkLockOnConflict(
 	fileLock metadata.FileLock,
 	fileID [16]byte,
 	lockSeqEnabled bool,
-	lockSeqIndex uint8,
-	lockSeqNumber uint32,
+	lockSeqIndex uint32,
+	lockSeqNumber uint8,
 ) (uint64, bool) {
 	// Deadlock detection: probe the conflicting holder via TestLock and
 	// check the WFG. If granting our wait would close a cycle, refuse to
